Migrate to LTS-7
Note that TypesF has really ugly instances because
TypeConstraintsF/ConstraintSetF is not supported by
the TH derivation logic of Data.Deriving.

It is unsupported because liftListEq/liftList are
rather non-trivial to implement efficiently (I ended
up using a non-efficient variant that scans lists
twice).

However this intermediate step lets us upgrade 5 of
the libraries which is better than trying to upgrade
all of them at once